requirements.txt
python-dotenv==1.2.1
groq==1.0.0
pyttsx3==2.99
sounddevice==0.5.3
scipy==1.16.3
soundfile==0.13.1
google-cloud-texttospeech==2.33.0
pvporcupine==4.0.1